irctc_backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0

mysqlclient==2.2.8
PyMySQL==1.1.1

firebase-admin==6.5.0

python-dotenv==1.0.1

gunicorn==22.0.0
whitenoise==6.7.0
reportlab==4.2.2
